+++ libs/vec3.h
@@ -2,33 +2,32 @@
 
 #include"commonLibs.h"
 
-
-
 class vec3 {
     public:
-    double x, y, z;
+    float x, y, z;
 
-    vec3(double x, double y, double z);
+    vec3(float x, float y, float z);
     vec3();
 
-    double lenth();
+    float lenth();
     bool operator==(const vec3 vector);
-    bool operator==(double len);
-    vec3 operator/(double coef);
-    vec3 operator*(double coef);
-    void operator*=(double coef);
-    void operator/=(double coef);
+    bool operator==(float len);
+    vec3 operator/(float coef);
+    vec3 operator*(float coef);
+    void operator*=(float coef);
+    void operator/=(float coef);
     vec3 operator+(const vec3 other_vector);
     void operator+=(const vec3 other_vector);
     vec3 operator-();
     vec3 operator-(const vec3 other_vector);
     void operator-=(const vec3 other_vector);
     vec3 normalize();
-    std::vector<double> getCords();    
+    std::vector<float> getCords();
+
 };
 
 std::ostream& operator<< (std::ostream& stream, const vec3& vec);
 
 vec3 vectorMult(vec3 vector1, vec3 vector2);
 
-double scalarMult(const vec3 vector1, const vec3 vector2);
+float scalarMult(const vec3 vector1, const vec3 vector2);
